updated checking for 401 and 403 status codes

+++ lib/screens/auth/reg/info/components/reg_info_page.dart
@@ -108,12 +108,11 @@ class _RegInfoFormPage extends State<RegInfoFormPage> {
                       await Connectivity().checkConnectivity();
 
                   if (connectivityResult != ConnectivityResult.none) {
-                    var statusCode;
-                    try{
-                       statusCode = await UserApi.updateUserInfo(
-                          firstName, secondName, gender, date, fileName);
-                    }catch(e){
-                      log(e.toString());
+                    var statusCode = await UserApi.updateUserInfo(
+                        firstName, secondName, gender, date, fileName);
+
+                    if (statusCode == HttpStatus.forbidden ||
+                        statusCode == HttpStatus.unauthorized) {
                       AuthUtils.deleteJwt();
                       Phoenix.rebirth(context);
                     }
@@ -131,7 +130,8 @@ class _RegInfoFormPage extends State<RegInfoFormPage> {
                       });
                     } else {
                       _btnController.error();
-                      AppUtils.displaySnackBar(context, "Ошибка добавления информации!");
+                      AppUtils.displaySnackBar(
+                          context, "Ошибка добавления информации!");
                       Timer(Duration(seconds: 1), () {
                         _btnController.reset();
                       });
